force the pos, vel and mass array type to be float in Particles

+++ Fireworks/fireworks/particles.py
@@ -71,17 +71,17 @@ class Particles:
         :param mass: A Nx1 numpy array containing the mass of the N particles
         """
 
-        self.pos = np.atleast_2d(position)
+        self.pos = np.atleast_2d(position, dtype=float)
         if self.pos.shape[1] != 3: print(f"Input position should contain a Nx3 array, current shape is {self.pos.shape}")
 
-        self.vel = np.atleast_2d(velocity)
+        self.vel = np.atleast_2d(velocity, dtype=float)
         if self.vel.shape[1] != 3: print(f"Input velocity should contain a Nx3 array, current shape is {self.pos.shape}")
         if len(self.vel) != len(self.pos): print(f"Position and velocity in input have not the same number of elemnts")
 
-        self.mass = np.atleast_1d(mass)
+        self.mass = np.atleast_1d(mass, dtype=float)
         if len(self.mass) != len(self.pos): print(f"Position and mass in input have not the same number of elemnts")
 
-        self.ID=np.arange(len(self.mass))
+        self.ID=np.arange(len(self.mass), dtype=int)
 
         self.acc=None
